column to module
