+++ polars/polars-lazy/src/physical_plan/expressions/column.rs
@@ -1,0 +1,31 @@
+use crate::physical_plan::state::ExecutionState;
+use crate::prelude::*;
+use polars_core::prelude::*;
+use std::sync::Arc;
+
+pub struct ColumnExpr(Arc<String>, Expr);
+
+impl ColumnExpr {
+    pub fn new(name: Arc<String>, expr: Expr) -> Self {
+        Self(name, expr)
+    }
+}
+
+impl PhysicalExpr for ColumnExpr {
+    fn as_expression(&self) -> &Expr {
+        &self.1
+    }
+    fn evaluate(&self, df: &DataFrame, _state: &ExecutionState) -> Result<Series> {
+        let column = match &**self.0 {
+            "" => df.select_at_idx(0).ok_or_else(|| {
+                PolarsError::NoData("could not select a column from an empty DataFrame".into())
+            })?,
+            _ => df.column(&self.0)?,
+        };
+        Ok(column.clone())
+    }
+    fn to_field(&self, input_schema: &Schema) -> Result<Field> {
+        let field = input_schema.field_with_name(&self.0).map(|f| f.clone())?;
+        Ok(field)
+    }
+}
