+++ polars/polars-lazy/src/physical_plan/expressions/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod binary;
+pub(crate) mod column;
 pub mod default;
 mod final_agg;
 pub(crate) mod literal;
